issue: 4207451 improve config system
Enhance configuration system robustness and maintainability:
 - Replace static err strs with ERROR_STRING macro for function context
 - json_object_handle class implementing RAII pattern for JSON object
 - Improve memory management and fix potential leaks in JSON processing
 - JSON schema with detailed titles and descriptions for all properties
 - Extract large functions into smaller, focused implementations
 - Improve type safety between int64_t and size_t
 - Rename error_format namespace to type_format for clarity
 - Refactor descriptor provider with cleaner error handling
 - Extract configuration profile settings into separate methods
 - Add README.md for new configuration system

+++ src/core/dev/allocator.cpp
@@ -504,7 +504,7 @@ repeat:
         data = (void *)((uintptr_t)m_blocks.back()->data() + m_latest_offset);
         m_latest_offset += actual_size;
     } else if (!m_b_hw) {
-        if (expand(std::max(static_cast<size_t>(safe_mce_sys().heap_metadata_block), actual_size))) {
+        if (expand(std::max(safe_mce_sys().heap_metadata_block, actual_size))) {
             goto repeat;
         }
     }
